# Minor png hashing optimizations