### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Minor optimization of png hashing methods changes scaled performance from `2.48` to `2.40` relative to System.Drawing

Method |      TestImage |     Mean |    Error |    StdDev | Scaled | ScaledSD |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
--------------------- |--------------- |---------:|---------:|----------:|-------:|---------:|--------:|--------:|--------:|----------:|
'System.Drawing Png' | Png/splash.png | 3.949 ms | 3.165 ms | 0.1788 ms |   1.00 |     0.00 | 70.3125 | 70.3125 | 70.3125 | 240.51 KB |
'ImageSharp Png' | Png/splash.png | 9.765 ms | 4.235 ms | 0.2393 ms |   2.48 |     0.11 | 46.8750 |       - |       - |  96.94 KB |

Method |      TestImage |     Mean |     Error |    StdDev | Scaled | ScaledSD |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
--------------------- |--------------- |---------:|----------:|----------:|-------:|---------:|--------:|--------:|--------:|----------:|
'System.Drawing Png' | Png/splash.png | 3.559 ms | 2.3499 ms | 0.1328 ms |   1.00 |     0.00 | 74.2188 | 74.2188 | 74.2188 | 240.54 KB |
'ImageSharp Png' | Png/splash.png | 8.549 ms | 0.5069 ms | 0.0286 ms |   2.40 |     0.08 | 46.8750 |       - |       - |  96.94 KB |

<!-- Thanks for contributing to ImageSharp! -->