@JimBobSquarePants 2017-12-06
@tocsoft Yeah, it's heavy going with the inlining but I think it's best to be safe than sorry here.

We need to be as fast as possible with these algorithms and their use is specific to png so code replication shouldn't be an issue. I'm hoping one day we can switch out the Crc32 method with the SIMD instruction once we have that API.